Initial commit to repo
The app is now working. To use it simply open it up via startWorktimeTracker.bat and the app will create a simple database file. All commands are displayed inside the app so there is no need to remember them. To clock out it's recommended to use the clockOutAndShutDown.bat file which does what it says it does.